+++ lib/msal-common/src/utils/Constants.ts
@@ -143,6 +143,7 @@ export enum AADServerParamKeys {
     FOCI = "foci",
     CCS_HEADER = "X-AnchorMailbox",
     RETURN_SPA_CODE = "return_spa_code",
+    NATIVE_BROKER = "nativebroker",
     LOGOUT_HINT = "logout_hint"
 }